requirements.txt
APScheduler==3.7.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
influxdb==5.3.1
msgpack==1.0.2
pycoingecko==1.4.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.3
